% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Adenoma-class}
\alias{Adenoma}
\alias{Adenoma-class}
\title{Class implementing Adenoma}
\description{
Class implementing Adenoma
}
\seealso{
Other CrcSpinModel_classes: \code{\link{AdenomaParams}},
  \code{\link{AdenomaParams-class}};
  \code{\link{CrcRiskParams}},
  \code{\link{CrcRiskParams-class}}; \code{\link{CrcRisk}},
  \code{\link{CrcRisk-class}}; \code{\link{CrcSpinModel}},
  \code{\link{CrcSpinModel-class}}; \code{\link{Risk}},
  \code{\link{Risk-class}}
}

