% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{AdenomaParams-class}
\alias{AdenomaParams}
\alias{AdenomaParams-class}
\title{Class that hold parameters used to derive factors used in model adenoma transitions}
\description{
This class contains a number of parameters utilised by \code{\link{Adenoma}}
to derive value for factors it uses to model the transition of Adenomas from
one state to the next (see \code{\link{Adenoma::transition}})
}
\details{
Upon object initialisation via \code{$new()} the parameters take on hard
default values (see \code{\link{initialize}} which can later be changed via
the \code{\link{setParams}} method.

Storing the parameters used to derive modeling values independent to the
\code{\link{Adenomas}} allows driver classes to set population wide modeling
parameters without needing to interface every Adenoma track in the model,
and to pass the parameters to other classes if required.
}
\section{Fields}{

\describe{
\item{\code{gamma1_min}}{}

\item{\code{gamma1_max}}{}

\item{\code{gamma2_min}}{}

\item{\code{gamma2_max}}{}

\item{\code{gamma3_val}}{}
}}
\seealso{
Other CrcSpinModel_classes: \code{\link{Adenoma}},
  \code{\link{Adenoma-class}}; \code{\link{CrcRiskParams}},
  \code{\link{CrcRiskParams-class}}; \code{\link{CrcRisk}},
  \code{\link{CrcRisk-class}}; \code{\link{CrcSpinModel}},
  \code{\link{CrcSpinModel-class}}; \code{\link{Risk}},
  \code{\link{Risk-class}}
}

