% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ClinicalHistory-class}
\alias{ClinicalHistory}
\alias{ClinicalHistory-class}
\title{Class implementing records of clinical history}
\description{
This class is used to keep a list of test and events which occur through a
subject's clinical history. The records are stored in a list field
\code{events} which can contain an object of virtually any kind or class. It
is the responsibility of the class or function using this clinical history
class to filter out \code{event} objects that they do not understand how to
interpret.
}
\details{
It is used by an associated \code{\link{Person}} class representing the
subjects studied in the \code{\link{GenericModel}} clinical model. However,
it aims is to act as generic event container that remains relevant across a
range of more complex clinical model derived from from those base classes.
}
\section{Fields}{

\describe{
\item{\code{status}}{.}

\item{\code{events}}{List of records on a subject's medical tests or events.}
}}
\seealso{
Other GenericModel_classes: \code{\link{GenericModel}},
  \code{\link{GenericModel-class}}; \code{\link{Person}},
  \code{\link{Person-class}}; \code{\link{Test}},
  \code{\link{Test-class}}
}

