% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{CrcRisk-class}
\alias{CrcRisk}
\alias{CrcRisk-class}
\title{Class representing a person's risk of developing adenomas that lead to CRC}
\description{
This class stores a person's CRC risk factors for use in the
\code{\link{CrcSpinModel}}, as well as a reference to the initial risk
modeling parameters used to generate the person's risk factors
}
\details{
In particular, it gives context to risk factors defined in \code{\link{Risk}}
(from which it derives) within the \code{\link{CrcSpinModel}} for a
\code{\link{PersonWithColon}} where it represents the risk of them
developing \code{\link{Adenoma}}s that lead to CRC developing.
}
\section{Fields}{

\describe{
\item{\code{risk_level}}{String "high"|"standard" that was specified for the
       associated person at initialisation}

\item{\code{crcrisk_model_params}}{Reference to the CrcRiskParams object specified
       at initialisation and used to seed/generate the risk factor values}

\item{\code{baseline_risk}}{Inherited from \code{\link{Risk}}}

\item{\code{sex_linked_risk}}{Inherited from \code{\link{Risk}}}

\item{\code{age_risk}}{A 4-vector representing risk at ages prior to 50, 50-60,
       60-70, 70+ respectively. Overwrites \code{\link{Risk}.age_risk}}
}}
\seealso{
Other CrcSpinModel_classes: \code{\link{AdenomaParams}},
  \code{\link{AdenomaParams-class}}; \code{\link{Adenoma}},
  \code{\link{Adenoma-class}}; \code{\link{CrcRiskParams}},
  \code{\link{CrcRiskParams-class}};
  \code{\link{CrcSpinModel}},
  \code{\link{CrcSpinModel-class}}; \code{\link{Risk}},
  \code{\link{Risk-class}}
}

