% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{CrcRiskParams-class}
\alias{CrcRiskParams}
\alias{CrcRiskParams-class}
\title{Class that hold parameters used to model a persons risk factors}
\description{
This class contains a number of parameters utilised by \code{\link{CrCRisk}}
to model/generate a person's risk factors, and by other classes in the
\code{\link{CrcSpinModel}} family.
}
\details{
Upon object initialisation via \code{$new()} the parameters take on hard
default values (see \code{\link{initialize}} which can later be changed via
the \code{\link{setParams}} method.

Storing the risk factor modeling parameters independent to the risk factor
class allows driver classes to set population wide modeling parameters
without needing to interface the risk factor class of every individual
directly, and to pass the parameters to other classes if required.
}
\section{Fields}{

\describe{
\item{\code{baseline_mean}}{Mean used to draw base line risk factor.}

\item{\code{baseline_sd}}{Standard deviation used to draw base line risk factor.}

\item{\code{sex_linked_mean}}{Mean used to draw gender linked risk factor.}

\item{\code{sex_linked_sd}}{Standard deviation used to draw gender linked risk factor.}

\item{\code{age_mean}}{Mean used to draw age risk factor.}

\item{\code{age_sd}}{Standard deviation used to draw age risk factor.}
}}
\seealso{
Other CrcSpinModel_classes: \code{\link{AdenomaParams}},
  \code{\link{AdenomaParams-class}}; \code{\link{Adenoma}},
  \code{\link{Adenoma-class}}; \code{\link{CrcRisk}},
  \code{\link{CrcRisk-class}}; \code{\link{CrcSpinModel}},
  \code{\link{CrcSpinModel-class}}; \code{\link{Risk}},
  \code{\link{Risk-class}}
}

