% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{death_rates}
\alias{death_rates}
\title{Death rates}
\format{two vectors of length 100}
\description{
A dataset containing Australain male and female death rates for 2008-2010, available from the
Australian Bureau of Statistics. See
\url{http://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/3302.0.55.001Main+Features12008-2010}
These death rate define a "period" life table, as though the whole population, of whatever age, was subject
to the current mortality rates. The alternative is a "cohort" life table that follows a cohort from
birth to age "x"  with the rates of martality experienced by that cohort.
}
\details{
\itemize{
  \item death_rate_female
  \item death_rate_male
}
}
\keyword{datasets}

