% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{DukesCrcSpinModel-class}
\alias{DukesCrcSpinModel}
\alias{DukesCrcSpinModel-class}
\title{Class orchestrating the Duke's CRC  Spin model}
\description{
This class acts as the entry point and driver for an implementation of the
Dukes CRC Spin model, described by X et al., which is implemented using a
collection of classes.
}
\details{
It orchestrates the modeling on a population represented as a
\code{study_group} of \code{\link{PersonWithColon}} class objects. And models
the development of \code{\link{Adenoma}}s, and ultimately CRC, within the
subjects' \code{\link{Colon}}.

Users of the model should only call this class's \code{new}
(\code{\link{initialize}}), \code{\link{set_adenoma_modeling_parameters}},
\code{\link{set_crcrisk_modeling_parameters}}, and \code{\link{run}}
methods. And directly access it's \code{\link{study_results field}}.
}
\section{Fields}{

\describe{
\item{\code{study_results}}{A two dimensional matrix containing an 18 column summary of model
       state at each year of the study. Inherited from \code{\link{GenericModel}}}
}}

