% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lappend}
\alias{lappend}
\title{append to a list}
\usage{
lappend(my.list, obj)
}
\arguments{
\item{my.list}{a list}

\item{obj}{an item to append to the list}
}
\description{
This function appends on item to a list
}
\examples{
my.list<-list("a","b","c")
lappend(my.list,"d")
}

