% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Risk-class}
\alias{Risk}
\alias{Risk-class}
\title{Class representing a person's risk from a number of factors}
\description{
This class represents a persons risk in some given context. It aims to be a
generic container of risk variable that can be reused in various models.
}
\section{Fields}{

\describe{
\item{\code{baseline_risk}}{}

\item{\code{sex_linked_risk}}{Additional risk factor based on a persons gender}

\item{\code{age_risk}}{Additional risk factor based on a persons current age}
}}
\seealso{
Other CrcSpinModel_classes: \code{\link{AdenomaParams}},
  \code{\link{AdenomaParams-class}}; \code{\link{Adenoma}},
  \code{\link{Adenoma-class}}; \code{\link{CrcRiskParams}},
  \code{\link{CrcRiskParams-class}}; \code{\link{CrcRisk}},
  \code{\link{CrcRisk-class}}; \code{\link{CrcSpinModel}},
  \code{\link{CrcSpinModel-class}}
}

