% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Test-class}
\alias{Test}
\alias{Test-class}
\title{Class implementing records of clinical tests}
\description{
Class implementing records of clinical tests
}
\section{Fields}{

\describe{
\item{\code{age}}{Age at which the test was performed.}

\item{\code{type}}{String identifier of the test.}

\item{\code{compliance}}{String describing subjects acceptance of the test.}

\item{\code{result}}{String describing the result of test.}

\item{\code{state}}{String describing the state of health the person is in at
                 time of test.}
}}
\seealso{
Other GenericModel_classes: \code{\link{ClinicalHistory}},
  \code{\link{ClinicalHistory-class}};
  \code{\link{GenericModel}},
  \code{\link{GenericModel-class}}; \code{\link{Person}},
  \code{\link{Person-class}}
}

